+++ src/modules/productModule/product.resolver.ts
@@ -15,8 +15,16 @@ import { VoteService } from '../vote/vote.service';
 import { CommentService } from '../comment/comment.service';
 import { Comment } from '../comment/comment.entitiy';
 import { Vote } from '../vote/vote.entity';
+import { CurrentUser } from '../user/decorator/user.decorator';
+import { User } from '../user/user.entity';
+import {
+	UnauthorizedException,
+	NotFoundException,
+	UseGuards
+} from '@nestjs/common';
+import { GraphQLAuth } from '../auth/guard/GqlAuth.guard';
 
-@Resolver(of => Product)
+@Resolver((of: any) => Product)
 export class ProductResolver {
 	constructor(
 		private readonly productService: ProductService,
@@ -31,43 +39,68 @@ export class ProductResolver {
 
 	@Query(returns => Product)
 	async product(
-		@Args({ name: 'id', type: () => ID }) id: string
+		@Args({ name: 'id', type: () => ID }) id: number
 	): Promise<Product | undefined> {
 		return await this.productService.findOne(id);
 	}
 
 	@Mutation(returns => Product)
+	@UseGuards(GraphQLAuth)
 	async addProduct(
 		@Args('newProduct') newProduct: newProductInput
 	): Promise<Product> {
 		return await this.productService.create(newProduct);
 	}
 
+	private async checkOwnership(
+		ownerId: number | undefined,
+		productId: number | undefined
+	) {
+		if (ownerId && productId) {
+			const product: Product | undefined = await this.productService.findOne(
+				productId
+			);
+			const isOwner = product?.makersIds.some(id => id == ownerId);
+			if (!isOwner) {
+				throw new UnauthorizedException("You can't alter what's not yours");
+			}
+		} else {
+			throw new NotFoundException('Please check your input');
+		}
+	}
+
 	@Mutation(returns => Product)
+	@UseGuards(GraphQLAuth)
 	async updateProduct(
-		@Args('updatedProduct') updatedProuduct: updatedProductInput
+		@Args('updatedProduct') updatedProuduct: updatedProductInput,
+		@CurrentUser() { id: userId }: User
 	): Promise<Product> {
-		return await this.productService.update(
-			updatedProuduct.id,
-			updatedProuduct
-		);
+		const { id: productId } = updatedProuduct;
+		await this.checkOwnership(userId, productId);
+
+		return await this.productService.update(productId, updatedProuduct);
 	}
 
 	@Mutation(returns => Boolean)
+	@UseGuards(GraphQLAuth)
 	async deleteProduct(
-		@Args({ name: 'id', type: () => ID }) id: string
+		@Args({ name: 'id', type: () => ID }) id: number,
+		@CurrentUser() { id: userId }: User
 	): Promise<Boolean> {
+		const product: Product | undefined = await this.productService.findOne(id);
+		const productId = product?.id;
+		await this.checkOwnership(userId, productId);
 		return await this.productService.delete(id);
 	}
 
 	@ResolveProperty('votes')
-	async votes(@Parent() product): Promise<Vote[]> {
+	async votes(@Parent() product: Product): Promise<Vote[]> {
 		const { id } = product;
 		return await this.voteService.findAll({ productId: id });
 	}
 
 	@ResolveProperty('comments')
-	async comments(@Parent() product): Promise<Comment[]> {
+	async comments(@Parent() product: Product): Promise<Comment[]> {
 		const { id } = product;
 		return await this.commentService.findAllComments(id);
 	}
